adding amount to the shortcode
I am developing a website and I use the version of timeline-express. I wanted to display the announcements on the home page, but not all. So I have edited the shortcode so that the user can give amount of announcements as a parameter :) 
I hope what I have done is okay ! 
Thanks